extra GreptConfig from Config
